Fix Select/Deselect all buttons in Medical Info section of Privacy Settings
Details: MedicalInformationModalController function in consentDirectives.js to manually set medicalInformationModalObj to 'dirty' state when Select All or Deselect All buttons in Medical Info section of Privacy Settings modal are clicked. This is necessary for the medicalInformationModalObj to recognize the form has been changed and enable the 'Save Changes' button in the Privacy Settings modal.

#7343-Create/Edit Consent :: Privacy Settings Modal :: Disable 'Save changes' Selection Until Change(s) Made